feat(nvim): enable processing of `.vscode/launch.json`
Enables the processing of `.vscode/launch.json` for debug
configurations. Had to add a en extensions specifically for processing
json5 files, which the launch.json is, to handle the extra commas and
inline comments that are prevalent in launch.json files. To make this
work with `nvim-dap`, I had to set this up in the `config()` function
for the `nvim-dap` plugin spec. LazyVim defines a `config()` function
already, so I added the necessary bits, then copied from
https://github.com/LazyVim/LazyVim/blob/81ab5bed7a81c3005f862d6508dacc26c4ed553a/lua/lazyvim/plugins/extras/dap/core.lua#L92-L101

Any time LzyVim is updated, need to verify this.